backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
sqlalchemy==2.0.36
openai==1.54.0
google-generativeai==0.8.3
google-cloud-vision==3.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
